Require `GFp_BN_mod_exp_mont_consttime()` inputs to be reduced.
Require that the input be reduced (mod n), simplifying the code.

Also, move the tests for `elem_exp_consttime()` to Rust and split the
vectors out into their own file, like was done for the
`elem_exp_vartime()` tests. The test vectors were taken directly from
BoringSSL, filtered as described in the comment at the top of the file.