Split internal storage of `Ed25519KeyPair`.
Before, we stored the bytes of the private and public key together in
one contiguous array to support using the foreign `GFp_ed25519_sign`
function when signing. After the port, we no longer use that function,
so we can switch to a more direct internal representation.

I agree to license my contributions to each file under the terms given
at the top of each file I changed.