Add methods for Curve25519 point encoding.
This safely wraps the unsafe C calls and gives us a destination for
their upcoming oxidization. We don't include a slice-mutating point
encoding method because its measured benefit was negligible.

Note that we stop using the `Scalar` alias, which was inaccurate.
Although `SCALAR_LEN` == `ELEM_LEN`, the byte arrays in question are
not encoded scalars, but "compressed" encodings of curve points. These
are 31-bit encodings of the Y coordinate of a point, enriched with a
sign bit. So, we change to `EncodedPoint`, which describes our actual
use of the type.

I agree to license my contributions to each file under the terms given
at the top of each file I changed.