Replace confusing use of `GFp_BN_mod_sub_quick()`.
`GFp_BN_mod_sub_quick()` is documented to require its non-modulus
operands to be less than the modulus, but in this case we were passing
the modulus itself as one of them. Avoid doing that, and add assertions
regarding the range of the inputs to `GFp_BN_mod_sub_quick()`.

Also, this makes it clearer that the function isn't constant-time yet.