Remove `BIGNUM::neg`: `BIGNUM`s are now always nonnegative/unsigned.
Note that in each place where the value of `neg` was changed for some
`BIGNUM`, either it is set to zero (not negative), the same value as
another `BIGNUM`'s `neg` value, or to the XOR of two `BIGNUM`'s `neg`
values. Thus, there was no way the `neg` flag could ever be set to a
nonzero value.