Add Debug implementation for Algorithms.
This changeset adds the ring::debug module, containing the
derive_debug_from_field!() macro, which generates an implementation of
fmt::Debug that defers the formatting to a given field.

This macro is then used to generate fmt::Debug implmentations for
aead::Algorithm and digest::Algorithm which print the algorithm id.

I agree to license my contributions to each file under the terms given
at the top of each file I changed.