Android CI: Fix the copying of the ring test executable.
Newer versions of rustc generate a "ring-$something.d" file that was
matched by the old pattern. Update the pattern to be more picky to
unbreak this part of the Android CI.

Now the Android stuff is at least built again. Running the tests in the
emulator still fails due to an unrelated issue.

Thanks to andrewtj for proposing this.