Fix AES on pre-AES-NI x86-64 systems.
In 0.13.0-alpha3 the BSAES implementation was removed. When that
happened, the VPAES implementation started getting used; previously it
was dead code. Long ago, the decryption code was removed from VPAES
(and all other AES implementations) since the modes we care about (CTR,
GCM, etc.) only uses the encryption primitive. When that old change was
made, this one instruction was removed when it shouldn't have been.

Thanks to Jan Andre Ikenmeyer for reporting this.