Add support for MIPS architectures.
This adds mips/mips64/mipsel/mips64el targets to the build system.
Unfortunately travis `trusty` image doesn't ship with `gcc-mips*`
packages. On these platforms `cargo` is replaced with `cross`[1] to run
those builds on a docker container.

For now All MIPS tests are allowed to fail to make way for future PRs
that will fill the gaps.

This is a first step to bring support to MIPS platforms.

[1]: https://github.com/japaric/cross