Expose unauthenticated stream cipher API
This adds a new "unauthenticated_stream" module, to expose a plain stream
cipher API, which mostly follows the aead module, but instead of providing
"open" and "seal" operations, it just provides a "xor_keystream" operation
and uses 128 bits nonces instead of 96 bits.

The 128 bits nonce are necessary to match the usage of the stream ciphers
for the QUIC protocol packet number protection scheme, where 16 octets are
sampled from the packet payload and used as explicit nonce for AES-CTR
and ChaCha20.

This also implements support for the ChaCha20 cipher, exposed through the
new API described above, by reusing the existing ChaCha20 C interface. The
first 32 bits of the 128 bits nonce are extracted and used as an explicit
initial counter.

The ChaCha20 test vectors were copied from the AEAD ChaCha20-Poly1305 ones,
with the nonces modified to add an explicit little-endian counter of "1", and
with AD and TAG parameters removed.

I agree to license my contributions to each file under the terms given
at the top of each file I changed.