Refactor C code for AES key expansion.
The control flow of the original code was confusing. This moves the
S-box computations into a separate function and simplifies the control
flow. The while loop for 256-bit key expansion remains the same, as it
facilitates loop unrolling.

I agree to license my contributions to each file under the terms given
at the top of each file I changed.