Remove 192-bit AES support in ARM assembly.
192-bit key support still exists in AES implementation for ARMv8.
Remove it, and leave comments documenting where code was removed.

As far as I can tell 192-bit keys are not supported in either the ARMv4
or vector-permutation based ARMv7 implementations.

This continues the work in
- 1103cf29dfbbf51f0dd8fb757084caa052863869
- b3e91be71edde28f5d2884d3c3c34260b6a79378
- and #707.

I agree to license my contributions to each file under the terms given
at the top of each file I changed.