Use Fn bound for kdf on agree_ephemeral.
The purpose of the kdf parameter would seem to be to restrict the use
of the key material.  The most restrictive function bound is Fn, which
requires that the kdf be read-only with respect to its environment.

I agree to license my contributions to each file under the terms given
at the top of each file I changed.