Add RSA private keys from components skeleton
This change prepares the path for the addition of an additional
constructor for rsa::KeyPair that takes in an RSA key's constituent
components. It introduces a new struct, KeyPairInput, that contains all
those components and adds the constructor's unimplemented skeleton.

I agree to license my contributions to each file under the terms given
at the top of each file I changed.