Make read_tag_and_get_value private to the module
This change makes read_tag_and_get_value private to the module it is
defined in. Since it is not used anywhere else, there is no need to have
it be public.

I agree to license my contributions to each file under the terms given
at the top of each file I changed.