Rename nonnegative_integer to read_nonnegative_integer
This change renames the nonnegative_integer function to
read_nonnegative_integer. For one, that name is more in line with
read_tag_and_get_value which accepts a reader and returns an
untrusted::Input, but it also helps prepares for introduction of a new
nonnegative_integer function that works on an untrusted::Input instead
of an untrusted::Reader.

I agree to license my contributions to each file under the terms
given at the top of each file I changed.