Introduce der::nonnegative_integer
This change introduces a new function, nonnegative_integer that accept
an untrusted::Input and verifies that it represents a non-negative
integer with a given value. The implementation really is just the guts
of read_nonnegative_integer.

I agree to license my contributions to each file under the terms
given at the top of each file I changed.