Provide an easier-to-use rand API
Create a new trait called AsBytesMut, which is used to bound the return
value of rand::value(), which creates fills a newly-created object with
random data. Because we don't have support for generic-sized arrays yet,
we use macros to provide impls for arrays of any size from 0 to 32,
inclusive.

I agree to license my contributions to each file under the terms given
at the top of each file I changed.