Remove pbkdf2_one_iteration test.
This test was part of the test that we handled zero iterations
correctly. Recently we switched to using `NonZero` which
prevents zero iterations from being requested at build time, so
that test was removed. Remove this leftover fragment too. Note that
there is still a test for one iteration in pkbdf2_tests.txt.