Unify multi-part signing and verification
This is not the same as the role fulfilled by the `SigningContext`. The
newly introduced API requires the complete message to be somewhere in
memory, it only does not require those chunks to be contiguous or in a
particular order. This prevents misuse of I-U-F (init-update-finalize)
style API for verification while aligning the available alternatives for
signing and verification. In particular, it is no longer necessary to
compare Tags directly for multi-part messages—provided the parts can be
referenced in memory simultaneously.

Closes: #615

I agree to license my contributions to each file under the terms given
at the top of each file I changed.