Use modern C initialization syntax
Closes #844

I reviewed the C code for how we handle uninitialized variables. It
looks like the uses of limbs_zero were the only times we immdiately
zero a value right after declaring it in `crypto/`.

We could switch some of the initializations in `third_party/fiat`,
but those are in the "Proven by Coq" section, so I'll just leave them
be.