Emit empty signerInfos in PKCS#7 bundles.
This is our bug that we've had since the beginning of PKCS#7 writing
support in eeb9f491: the empty signerInfos SET wasn't emitted. Some
parsers, including OpenSSL, don't like this but it appears to have taken
five years for anyone to notice.

This change does not make parsing strict so that we continue to parse
old messages that we may have produced.

(As ever, PKCS#* should not be used expect where absolutely required for
interoperability.)

Bug: b:135982177

Change-Id: Ia7241de69f105657bdfb5ff75e909deae71748a0
Reviewed-on: https://boringssl-review.googlesource.com/c/boringssl/+/36564
Commit-Queue: David Benjamin <davidben@google.com>
Reviewed-by: David Benjamin <davidben@google.com>