ChaCha20-Poly1305: Define integrated assembly types more like BoringSSL.
Define the input data structures for the integrated assembly almost exactly
like BoringSSL does, except for the caveat mentioned in the comments.

Similarly, define the output type as a structure containing a byte array, like
BoringSSL, again with the caveats mentioned in the comments.

Abstract the union types into a single parameterized type.