ChaCha: Reduce the use of `unsafe`; emulate the `copy_within` API.
Emulate the `copy_within` API so that the use of `unsafe` can be isolated
into a single place.

Remove the test for encryption into disjoint buffers. We should expose
such an API in the future, but currently we don't, so this was testing a
scenerio that would never occur. Removing this part of the test was
necessary to enable this refactoring. We'll need to bring it back when
we implement out-of-place encryption.