CI/CD: Remove optimization for repository owner's PRs.
The optimization avoids every CI job running twice for the repository owner's
PRs. This is a pretty useful optimization because there are so many slow jobs
that are run for every commit that are duplicated for each of my PRs. However,
in the interest of making CI as bulletproof as possible, I'm going to try to
eliminate this. We'll add it back, or similar, if things get too bad.

```
$ grep "if:" .github/workflows/ci.yml | sort | uniq
      - if: ${{ !contains(matrix.host_os, 'windows') }}
      - if: ${{ contains(matrix.host_os, 'ubuntu') }}
      - if: ${{ contains(matrix.host_os, 'windows') }}
      - if: ${{ matrix.target == 'aarch64-apple-darwin' }}
```